requirements/tests.txt
-r app.txt
isort==4.2.2
pep257==0.7.0
pep8==1.7.0
pylint==1.5.4
pylint-django==0.7.1
pytest==2.8.7
pytest-django==2.9.1
pytest-pylint==0.4.1
